Initialize locks lazily
It seems that during class creation time, the event loop that is automatically
detected by asyncio.Lock is not the same event loop that ends up being used
by asyncio.run. Therefore, when two coroutines try to access the lock at the
same time and one of them needs to await it, a RuntimeError is raised because
of mismatched event loops.